# apps/mrerp/backend/openapi.yaml
openapi: 3.0.3
info:
  title: MRERP API
  version: 1.0.0
  description: MRERP modular-monolith API contract
paths:
...
  /api/v1/leave/attendance/adjust/:
    post:
      operationId: api_v1_leave_attendance_adjust_create
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttendanceAdjustment'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AttendanceAdjustment'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AttendanceAdjustment'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceRow'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: ''
  /api/v1/leave/requests/:
    get:
      operationId: api_v1_leave_requests_list
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      tags:
      - api
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedLeaveRequestList'
          description: ''
    post:
      operationId: api_v1_leave_requests_create
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeaveRequestCreate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/LeaveRequestCreate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/LeaveRequestCreate'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaveRequest'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: ''
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        code:
          type: string
        detail:
          type: string
        correlation_id:
          type: string
          format: uuid
        errors:
          type: object
          additionalProperties: {}
      required:
      - code
      - correlation_id
      - detail
...
    AttendanceAdjustment:
      type: object
      properties:
        employee_uuid:
          type: string
          format: uuid
        month:
          type: string
          pattern: ^\d{4}-\d{2}$
        days:
          type: integer
          maximum: 31
          minimum: -31
        reason:
          type: string
          maxLength: 500
          minLength: 1
      required:
      - days
      - employee_uuid
      - month
      - reason
...
    AttendanceRow:
      type: object
      properties:
        employee_uuid:
          type: string
          format: uuid
        employee_code:
          type: string
        display_name:
          type: string
        team_name:
          type: string
          nullable: true
        month:
          type: string
        scheduled_workdays:
          type: number
          format: double
        public_holiday_days:
          type: number
          format: double
        approved_leave_days:
          type: number
          format: double
        adjustment_days:
          type: integer
        adjustment_reason:
          type: string
        projected_workdays:
          type: number
          format: double
      required:
      - adjustment_days
      - adjustment_reason
      - approved_leave_days
      - display_name
      - employee_code
      - employee_uuid
      - month
      - projected_workdays
      - public_holiday_days
      - scheduled_workdays
      - team_name
...
    EndPeriod457Enum:
      enum:
      - am
      - pm
      type: string
      description: |-
        * `am` - Sáng
        * `pm` - Chiều
    EndPeriod4efEnum:
      enum:
      - am
      - pm
      type: string
      description: |-
        * `am` - am
        * `pm` - pm
...
    LeaveRequest:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
          readOnly: true
        requester_code:
          type: string
          readOnly: true
        requester_name:
          type: string
          readOnly: true
        team_name:
          type: string
          readOnly: true
          nullable: true
        start_date:
          type: string
          format: date
          readOnly: true
        end_date:
          type: string
          format: date
          readOnly: true
        start_period:
          allOf:
          - $ref: '#/components/schemas/StartPeriod457Enum'
          readOnly: true
        end_period:
          allOf:
          - $ref: '#/components/schemas/EndPeriod457Enum'
          readOnly: true
        reason:
          type: string
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/LeaveRequestStatusEnum'
          readOnly: true
        status_label:
          type: string
          readOnly: true
        reviewer_name:
          type: string
          readOnly: true
          nullable: true
        review_note:
          type: string
          readOnly: true
        reviewed_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        version:
          type: integer
          readOnly: true
        can_review:
          type: boolean
          readOnly: true
        can_edit:
          type: boolean
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - can_edit
      - can_review
      - created_at
      - end_date
      - end_period
      - reason
      - requester_code
      - requester_name
      - review_note
      - reviewed_at
      - reviewer_name
      - start_date
      - start_period
      - status
      - status_label
      - team_name
      - updated_at
      - uuid
      - version
    LeaveRequestCreate:
      type: object
      properties:
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        start_period:
          allOf:
          - $ref: '#/components/schemas/StartPeriod4efEnum'
          default: am
        end_period:
          allOf:
          - $ref: '#/components/schemas/EndPeriod4efEnum'
          default: pm
        reason:
          type: string
          maxLength: 2000
          minLength: 1
      required:
      - end_date
      - reason
      - start_date
    LeaveRequestStatusEnum:
      enum:
      - pending
      - approved
      - rejected
      type: string
      description: |-
        * `pending` - Chờ duyệt
        * `approved` - Đã duyệt
        * `rejected` - Từ chối
...
    PaginatedLeaveRequestList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/LeaveRequest'
...
    StartPeriod457Enum:
      enum:
      - am
      - pm
      type: string
      description: |-
        * `am` - Sáng
        * `pm` - Chiều
    StartPeriod4efEnum:
      enum:
      - am
      - pm
      type: string
      description: |-
        * `am` - am
        * `pm` - pm
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid